+++ Tests/GuestConfigurationModule.Tests.ps1
@@ -39,7 +39,6 @@ function Get-OSPlatform {
 $IsNotAzureDevOps = $false -eq (Get-IsAzureDevOps)
 $IsNotWindows = 'Windows' -ne (Get-OSPlatform)
 $IsNotWindowsAndIsAzureDevOps = $IsNotWindows -AND (Get-IsAzureDevOps)
-$IsPester4 = (Get-Module 'Pester').version.Major -eq '4'
 
 if ($Env:BUILD_DEFINITIONNAME -eq 'PowerShell.GuestConfiguration (Private)') {
     $releaseBuild = $true
@@ -118,7 +117,7 @@ Import-Certificate -FilePath "$TestDrive/exported.cer" -CertStoreLocation Cert:\
 '@                
             powershell.exe -NoProfile -NonInteractive -Command $command
         }
-
+        
         function New-TestDscConfiguration {
             [CmdletBinding()]
             param
@@ -126,19 +125,17 @@ Import-Certificate -FilePath "$TestDrive/exported.cer" -CertStoreLocation Cert:\
                 [Parameter(Mandatory = $true)]
                 [String]
                 $DestinationFolderPath,
-
+        
                 [Parameter()]
-                [ValidateSet('DSC', 'InSpec')]
+                [ValidateSet('DSC', 'InSpec', 'WinDSC')]
                 [String]
                 $Type = 'DSC'
             )
-
+        
             if ($false -eq (Test-CurrentMachineIsWindows)) {
                 Import-Module 'PSDesiredStateConfiguration'
             }
-
-            Install-Module -Name 'ComputerManagementDsc' -AllowClobber -Force
-
+        
             #region Windows DSC config
             if ('DSC' -eq $Type) {
                 Install-Module -Name 'ComputerManagementDsc' -RequiredVersion '8.2.0' -Force
@@ -162,26 +159,40 @@ CompatibleVersionAdditionalProperties= {"Omi_BaseResource:ConfigurationName"};
 Name="DSCConfig";
 };
 '@
+
+            $dscDestinationFolderPath = New-Item -Path $DestinationFolderPath -Name 'DSCConfig' -ItemType Directory
+            $dscDestinationMOFPath = Join-Path -Path $dscDestinationFolderPath -ChildPath 'localhost.mof'
+            $null = Set-Content -Path $dscDestinationMOFPath -Value $dscConfig
+            
+            $filesToIncludeFolderPath = Join-Path -Path $DestinationFolderPath -ChildPath 'FilesToInclude'
+            New-Item $filesToIncludeFolderPath -ItemType Directory
+            $filesToIncludeFilePath = Join-Path -Path $filesToIncludeFolderPath -ChildPath 'file.txt'
+            $filesToIncludeContent = 'test' | Set-Content -Path $filesToIncludeFilePath
             }
             #endregion
         
             #region Linux DSC config
-            if ('InSpec' -eq $Type) {
-                $dscConfig = @"
-Configuration DSCConfig
+            if ('Inspec' -eq $Type) {
+                $dscConfig = @'
+instance of MSFT_ChefInSpecResource as $MSFT_ChefInSpecResource1ref
 {
-Import-DscResource -ModuleName 'GuestConfiguration'
+Name = "linux-path";
+ResourceID = "[ChefInSpecResource]Audit Linux path exists";
+ModuleVersion = "2.0.0";
+SourceInfo = "::7::1::ChefInSpecResource";
+ModuleName = "GuestConfiguration";
+ConfigurationName = "DSCConfig";
+};
 
-Node 'localhost'
+instance of OMI_ConfigurationDocument
 {
-ChefInSpecResource 'Audit Linux path exists'
-{
-    Name = 'linux-path'
-}
-}
-}
-DSCConfig -OutputPath $DestinationFolderPath
-"@
+Version="2.0.0";
+MinimumCompatibleVersion = "1.0.0";
+CompatibleVersionAdditionalProperties= {"Omi_BaseResource:ConfigurationName"};
+Name="DSCConfig";
+};                
+'@
+
                 $inSpecProfileName = 'linux-path'
                 $inSpecProfile = @"
 name: $inSpecProfileName
@@ -198,38 +209,55 @@ describe file('/tmp') do
 it { should exist }
 end
 "@
-            }
-            #endregion
-        
-            $DestinationFolderPath = New-Item -Path $TestDrive -Name 'DSCConfig' -ItemType Directory
-            $destinationMOFPath = Join-Path -Path $DestinationFolderPath -ChildPath 'localhost.mof'
-        
-            $null = Set-Content -Path $destinationMOFPath -Value $dscConfig
+                $inspecDestinationFolderPath = New-Item -Path $DestinationFolderPath -Name 'InspecConfig' -ItemType Directory
+                $inspecDestinationMOFPath = Join-Path -Path $inspecDestinationFolderPath -ChildPath 'localhost.mof'
+                $null = Set-Content -Path $inspecDestinationMOFPath -Value $dscConfig    
 
-            $filesToIncludeFolderPath = Join-Path -Path $TestDrive -ChildPath 'FilesToInclude'
-            New-Item $filesToIncludeFolderPath -ItemType Directory
-            $filesToIncludeFilePath = Join-Path -Path $filesToIncludeFolderPath -ChildPath 'file.txt'
-            $filesToIncludeContent = 'test' | Set-Content -Path $filesToIncludeFilePath
-        
-            if ('InSpec' -eq $Type) {
-                # creates directory for InSpec profile
-                $InSpecProfilePath = Join-Path -Path $TestDrive -ChildPath $inSpecProfileName
+                # creates directory for Inspec profile
+                $InSpecProfilePath = Join-Path -Path $inspecDestinationFolderPath -ChildPath $inSpecProfileName
                 $null = New-Item -ItemType Directory -Path $InSpecProfilePath
         
-                # creates InSpec profile required Yml file
-                $InSpecProfileYmlFilePath = Join-Path -Path $InSpecProfilePath -ChildPath 'inspec.yml'
+                # creates Inspec profile required Yml file
+                $InSpecProfileYmlFilePath = Join-Path -Path $InSpecProfilePath -ChildPath 'Inspec.yml'
                 $null = Set-Content -Path $InSpecProfileYmlFilePath -Value $inSpecProfile
         
-                # creates directory for InSpec controls (component of InSpec profile)
+                # creates directory for Inspec controls (component of Inspec profile)
                 $InSpecControlsPath = Join-Path -Path $InSpecProfilePath -ChildPath 'controls'
                 $null = New-Item -ItemType Directory -Path $InSpecControlsPath
         
-                # creates InSpec controls required Ruby file
+                # creates Inspec controls required Ruby file
                 $InSpecControlsRubyFilePath = Join-Path -Path $InSpecControlsPath -ChildPath "$inSpecProfileName.rb"
                 $null = Set-Content -Path $InSpecControlsRubyFilePath -Value $inSpecProfileRB
             }
+            #endregion 
+
+            #region Windows DSC config using invalid resources
+            if ('WinDSC' -eq $Type) {
+                $dscConfig = @'
+instance of MSFT_FileDirectoryConfiguration as $MSFT_FileDirectoryConfiguration1ref
+{
+ResourceID = "[File]test";
+Ensure = "Present";
+Contents = "test";
+DestinationPath = "c:\\test";
+ModuleName = "PSDesiredStateConfiguration";
+SourceInfo = "::1::76::file";
+ModuleVersion = "1.0";
+ConfigurationName = "file";
+};
+
+instance of OMI_ConfigurationDocument
+{
+Version="2.0.0";
+MinimumCompatibleVersion = "1.0.0";
+CompatibleVersionAdditionalProperties= {"Omi_BaseResource:ConfigurationName"};
+Name="DSCConfig";
+};
+'@
+            }
+            #endregion
         }
-        
+    
         function New-TestGCPolicyParameters {
             [CmdletBinding()]
             param
@@ -249,13 +277,51 @@ end
             
             $newGCPolicyParameters = @{
                 ContentUri  = 'https://github.com/microsoft/PowerShell-DSC-for-Linux/raw/amits/custompolicy/new_gc_policy/AuditWindowsService.zip'
-                DisplayName = "[Test] Audit Windows Service - Date: $currentDateString OS: $currentWindowsOSString"
+                DisplayName = "[ModuleTestCI] Audit Windows Service - Date: $currentDateString OS: $currentWindowsOSString"
                 Description = 'Policy to audit a Windows service'
                 Path        = Join-Path -Path $DestinationFolderPath -ChildPath 'policyDefinitions'
                 Version     = '1.0.0.0'
             }
         
             return $newGCPolicyParameters
+        }
+
+        function New-PublishGCPackageParameters {
+            [CmdletBinding()]
+            param
+            (
+                [Parameter(Mandatory = $true)]
+                [String]
+                $Path,
+
+                [Parameter(Mandatory = $true)]
+                [String]
+                $DateStamp
+            )
+
+            # Create test Resource Group
+            $resourceGroup = New-AzResourceGroup "GC_Module_$DateStamp" -Location 'westus'
+
+            # Create test Storage Account
+            $randomString = (get-date).ticks.tostring().Substring(12)
+            $storageAccount = New-AzStorageAccount -ResourceGroupName $resourceGroup.ResourceGroupName `
+                -Name "sa$randomString" `
+                -SkuName Standard_LRS `
+                -Location 'westus'
+
+            $ctx = $storageAccount.Context
+
+            # Storage Container
+            $containerName = "guestconfiguration"
+            New-AzStorageContainer -Name $containerName -Context $ctx -Permission blob
+            
+            $publishGCPackageParameters = @{
+                Path                 = $Path
+                ResourceGroupName    = "GC_Module_$DateStamp"
+                StorageAccountName   = "sa$randomString"
+            }
+        
+            return $publishGCPackageParameters
         }
 
         function Get-AzMocks {
@@ -349,26 +415,36 @@ end
     
         # Set up test paths
         $dscConfigFolderPath = Join-Path -Path $TestDrive -ChildPath 'DSCConfig'
-        $filesToIncludeFolderPath = Join-Path -Path $TestDrive -ChildPath 'FilesToInclude'
+        $mofPath = Join-Path -Path $dscConfigFolderPath -ChildPath 'localhost.mof'
         $testOutputPath = Join-Path -Path $TestDrive -ChildPath 'output'
-        $newPolicyDirectory = Join-Path -Path $testOutputPath -ChildPath 'policyDefinitions'
+        $newPolicyDirectory = Join-Path -Path $testOutputPath -ChildPath 'PolicyDefinitions'
         $policyName = 'testPolicy'
-        $mofDocPath = Join-Path -Path $dscConfigFolderPath -ChildPath 'localhost.mof'
-        $testPackagePath = Join-Path -Path $testOutputPath -ChildPath 'package'
-        $filesToIncludePackagePath = Join-Path -Path $testOutputPath -ChildPath 'filesToIncludePackage'
+        $testPackagePath = Join-Path -Path $testOutputPath -ChildPath 'Package'
         $unsignedPackageExtractionPath = Join-Path $testOutputPath -ChildPath 'UnsignedPackage'
-        $filesToIncludeExtractionPath = Join-Path $testOutputPath -ChildPath 'filesToIncludeUnsignedPackage'
-        $mofFilePath = Join-Path -Path $unsignedPackageExtractionPath -ChildPath "$policyName.mof"
+        $filesToIncludeFolderPath = Join-Path -Path $TestDrive -ChildPath 'FilesToInclude'
+        $filesToIncludePackagePath = Join-Path -Path $testOutputPath -ChildPath 'FilesToIncludePackage'
+        $filesToIncludeExtractionPath = Join-Path $testOutputPath -ChildPath 'FilesToIncludeUnsignedPackage'
         $extractedFilesToIncludePath = Join-Path -Path $filesToIncludeExtractionPath -ChildPath 'FilesToInclude'
+        $mofFilePath = Join-Path -Path $unsignedPackageExtractionPath -ChildPath "$policyName.mof"
+        $inSpecFolderPath = Join-Path -Path $TestDrive -ChildPath 'InspecConfig'
+        $inspecMofPath = Join-Path -Path $inSpecFolderPath -ChildPath 'localhost.mof'
+        $inspecPackagePath = Join-Path -Path $testOutputPath -ChildPath 'InspecPackage'
+        $inspecExtractionPath = Join-Path $testOutputPath -ChildPath 'InspecUnsignedPackage'
+        $inspecProfileName = 'linux-path'
+        $extractedInSpecPath = Join-Path -Path $inspecExtractionPath -ChildPath (Join-Path 'Modules' $inspecProfileName)
         $signedPackageExtractionPath = Join-Path $testOutputPath -ChildPath 'SignedPackage'
         $currentDateString = Get-Date -Format "yyyy-MM-dd HH:mm"
         $expectedPolicyType = 'Custom'
         $expectedContentHash = 'D421E3C8BB2298AEC5CFD95607B91241B7D5A2C88D54262ED304CA1FD01370F3'
         $testPolicyName = 'AuditWindowsService'
+
+        $Date = Get-Date
+        $DateStamp = "$($Date.Hour)_$($Date.Minute)_$($Date.Second)_$($Date.Month)-$($Date.Day)-$($Date.Year)"
         
         $newGCPolicyParameters = New-TestGCPolicyParameters $testOutputPath
-
-        New-TestDscConfiguration -DestinationFolderPath $dscConfigFolderPath
+        
+        New-TestDscConfiguration -DestinationFolderPath $TestDrive
+        New-TestDscConfiguration -DestinationFolderPath $TestDrive -Type 'Inspec'
 
         if ($Env:BUILD_DEFINITIONNAME -eq 'PowerShell.GuestConfiguration (Private)' -AND $false -eq $IsMacOS) {
             # TODO
@@ -394,7 +470,7 @@ end
             # TODO
             # Az PowerShell login from macOS currently has issue
             # https://github.com/microsoft/azure-pipelines-tasks/issues/12030
-            Install-AzLibraries
+            Install-AzLibraries            
         }
     }
     Context 'Module fundamentals' {
@@ -426,8 +502,8 @@ end
     }
     Context 'New-GuestConfigurationPackage' {
 
-        It 'creates Custom policy package' {
-            $package = New-GuestConfigurationPackage -Configuration $mofDocPath -Name $policyName -Path $testPackagePath
+        It 'creates custom policy package' {
+            $package = New-GuestConfigurationPackage -Configuration $mofPath -Name $policyName -Path $testPackagePath
             Test-Path -Path $package.Path | Should -BeTrue
             $package.Name | Should -Be $policyName
         }
@@ -439,7 +515,6 @@ end
             $null = Add-Type -AssemblyName System.IO.Compression.FileSystem
             { [System.IO.Compression.ZipFile]::ExtractToDirectory($package.FullName, $unsignedPackageExtractionPath) } | Should -Not -Throw
         }
-  
         It 'Verify extracted mof document exists' {
             Test-Path -Path $mofFilePath | Should -BeTrue
         }
@@ -461,7 +536,7 @@ end
         }
 
         It 'Implements -FilesToInclude parameter' {
-            $package = New-GuestConfigurationPackage -Configuration $mofDocPath -Name $policyName -Path $filesToIncludePackagePath -FilesToInclude $FilesToIncludeFolderPath
+            $package = New-GuestConfigurationPackage -Configuration $mofPath -Name $policyName -Path $filesToIncludePackagePath -FilesToInclude $FilesToIncludeFolderPath
             $null = Add-Type -AssemblyName System.IO.Compression.FileSystem
             { [System.IO.Compression.ZipFile]::ExtractToDirectory($package.Path, $filesToIncludeExtractionPath) } | Should -Not -Throw
             Test-Path -Path $extractedFilesToIncludePath | Should -BeTrue
@@ -469,11 +544,24 @@ end
             Test-Path -Path $extractedFile | Should -BeTrue
             Get-Content $extractedFile | Should -Be 'test'
         }
+
+        It 'Implements -ChefInspecProfilePath parameter' {
+            $package = New-GuestConfigurationPackage -Configuration $inspecMofPath -Name $policyName -Path $inspecPackagePath -ChefInspecProfilePath $inSpecFolderPath
+            $null = Add-Type -AssemblyName System.IO.Compression.FileSystem
+            { [System.IO.Compression.ZipFile]::ExtractToDirectory($package.Path, $inspecExtractionPath) } | Should -Not -Throw
+            $extractedInspecPath | Should -Exist
+            $inspecYmlExtractedFile = Join-Path $extractedInspecPath 'Inspec.yml'
+            $inspecYmlExtractedFile | Should -Exist
+            $inspecControlsExtractedFile = Join-Path $extractedInspecPath 'controls'
+            $inspecControlsExtractedFile | Should -Exist
+            $inspecRbExtractedFile = Join-Path $inspecControlsExtractedFile 'linux-path.rb'
+            $inspecRbExtractedFile | Should -Exist
+        }
     }
     Context 'Test-GuestConfigurationPackage' {
 
         It 'Validate that the resource compliance results are as expected' -Skip:$IsNotWindows {
-            $package = New-GuestConfigurationPackage -Configuration $mofDocPath -Name $policyName -Path $testPackagePath
+            $package = New-GuestConfigurationPackage -Configuration $mofPath -Name $policyName -Path $testPackagePath
             $testPackageResult = Test-GuestConfigurationPackage -Path $package.Path
             $testPackageResult.complianceStatus | Should -Be $false
             $testPackageResult.resources[0].ModuleName | Should -Be 'ComputerManagementDsc'
@@ -485,7 +573,7 @@ end
     Context 'Protect-GuestConfigurationPackage' {
         
         It 'Signed package should exist at output path' -Skip:$IsNotWindows {
-            $package = New-GuestConfigurationPackage -Configuration $mofDocPath -Name $policyName -Path $testPackagePath
+            $package = New-GuestConfigurationPackage -Configuration $mofPath -Name $policyName -Path $testPackagePath
             New-TestCertificate
             $certificatePath = "Cert:\LocalMachine\My"
             $certificate = Get-ChildItem -Path $certificatePath | Where-Object { ($_.Subject -eq "CN=testcert") } | Select-Object -First 1
@@ -514,9 +602,22 @@ end
             $authenticodeSignature.SignerCertificate.Thumbprint | Should -Be $certificate.Thumbprint
         }
     }
+    Context 'Publish-GuestConfigurationPackage' {
+
+        It 'Should be able to publish packages and return a valid Uri' -Skip:($notReleaseBuild -or $IsNotWindowsAndIsAzureDevOps) {
+            Login-ToTestAzAccount
+            $package = New-GuestConfigurationPackage -Configuration $mofPath -Name $policyName -Path $testPackagePath
+            $publishGCPackageParameters = New-PublishGCPackageParameters -Path $package.Path -DateStamp $DateStamp
+            $Uri = Publish-GuestConfigurationPackage -Path $publishGCPackageParameters.Path -ResourceGroupName $publishGCPackageParameters.ResourceGroupName -StorageAccountName $publishGCPackageParameters.StorageAccountName
+            $Uri | Should -Not -BeNullOrEmpty
+            $Uri | Should -BeOfType 'String'
+            $Uri | Should -Not -Contain '@'
+            {Invoke-WebRequest -Uri $Uri -OutFile $TestDrive/downloadedPackage.zip} | Should -Not -Throw
+        }
+    }
     Context 'New-GuestConfigurationPolicy' {
 
-        It 'New-GuestConfigurationPolicy should output path to generated policies' -Skip:($IsPester4 -or $IsNotWindowsAndIsAzureDevOps) {
+        It 'New-GuestConfigurationPolicy should output path to generated policies' -Skip:($IsNotWindowsAndIsAzureDevOps) {
             if ($notReleaseBuild) {
                 function Get-AzContext {}
                 Get-AzMocks -newGCPolicyParameters $newGCPolicyParameters
@@ -529,71 +630,50 @@ end
             Test-Path -Path $newGCPolicyResult.Path | Should -BeTrue
         }
 
-        It 'Generated Audit policy file should exist' -Skip:($IsPester4 -or $IsNotWindowsAndIsAzureDevOps) {
+        It 'Generated Audit policy file should exist' -Skip:($IsNotWindowsAndIsAzureDevOps) {
             $auditPolicyFile = Join-Path -Path $newPolicyDirectory -ChildPath 'AuditIfNotExists.json'
             Test-Path -Path $auditPolicyFile | Should -BeTrue
         }
 
-        It 'Audit policy should contain expected content' -Skip:($IsPester4 -or $IsNotWindowsAndIsAzureDevOps) {
+        It 'Audit policy should contain expected content' -Skip:($IsNotWindowsAndIsAzureDevOps) {
             $auditPolicyFile = Join-Path -Path $newPolicyDirectory -ChildPath 'AuditIfNotExists.json'
             $auditPolicyContent = Get-Content $auditPolicyFile | ConvertFrom-Json | ForEach-Object { $_ }
             $auditPolicyContent.properties.displayName.Contains($newGCPolicyParameters.DisplayName) | Should -BeTrue
             $auditPolicyContent.properties.description.Contains($newGCPolicyParameters.Description) | Should -BeTrue
+            $auditPolicyContent.properties.parameters.IncludeArcMachines | Should -Not -BeNullOrEmpty
             $auditPolicyContent.properties.policyType | Should -Be $expectedPolicyType
             $auditPolicyContent.properties.policyRule.then.details.name | Should -Be $testPolicyName
-        }
-
-        It 'Generated Deploy policy file should exist' -Skip:($IsPester4 -or $IsNotWindowsAndIsAzureDevOps) {
-            $deployPolicyFile = Join-Path -Path $newPolicyDirectory -ChildPath 'DeployIfNotExists.json'            
-            Test-Path -Path $deployPolicyFile | Should -BeTrue
-        }
-
-        It 'Deploy policy should contain expected content' -Skip:($IsPester4 -or $IsNotWindowsAndIsAzureDevOps) {
-            $deployPolicyFile = Join-Path -Path $newPolicyDirectory -ChildPath 'DeployIfNotExists.json'
-            $deployPolicyContent = Get-Content $deployPolicyFile | ConvertFrom-Json | ForEach-Object { $_ }
-            $deployPolicyContent.properties.displayName.Contains($newGCPolicyParameters.DisplayName) | Should -BeTrue
-            $deployPolicyContent.properties.description.Contains($newGCPolicyParameters.Description) | Should -BeTrue
-            $deployPolicyContent.properties.policyType | Should -Be $expectedPolicyType
-            $deployPolicyContent.properties.policyRule.then.details.deployment.properties.parameters.configurationName.value | Should -Be $testPolicyName
-            $deployPolicyContent.properties.policyRule.then.details.deployment.properties.parameters.contentHash.value | Should -Be $expectedContentHash
-            $deployPolicyContent.properties.policyRule.then.details.deployment.properties.parameters.contentUri.value | Should -Be $newGCPolicyParameters.ContentUri
         }
     }
     Context 'Publish-GuestConfigurationPolicy' {
 
-        It 'Should be able to publish policies' -Skip:($IsPester4 -or $notReleaseBuild -or $IsNotWindowsAndIsAzureDevOps) {
+        It 'Should be able to publish policies' -Skip:($notReleaseBuild -or $IsNotWindowsAndIsAzureDevOps) {
             Login-ToTestAzAccount
             $newGCPolicyResult = New-GuestConfigurationPolicy @newGCPolicyParameters
             { $publishGCPolicyResult = $newGCPolicyResult | Publish-GuestConfigurationPolicy } | Should -Not -Throw
         }
 
-        It 'Should be able to retrieve 2 published policies' -Skip:($IsPester4 -or $notReleaseBuild -or $IsNotWindowsAndIsAzureDevOps) {
+        It 'Should be able to retrieve 1 published policies' -Skip:($notReleaseBuild -or $IsNotWindowsAndIsAzureDevOps) {
             Login-ToTestAzAccount
             $existingPolicies = @(Get-AzPolicyDefinition | Where-Object { ($_.Properties.PSObject.Properties.Name -contains 'displayName') -and ($_.Properties.displayName.Contains($newGCPolicyParameters.DisplayName) ) } )
             write-host $($existingPolicies | % Properties)
             $null -ne $existingPolicies | Should -BeTrue
-            $existingPolicies.Count | Should -Be 2
-        }
-        
-        It 'Should be able to retrieve 1 published initiative' -Skip:($IsPester4 -or $notReleaseBuild -or $IsNotWindowsAndIsAzureDevOps) {
-            Login-ToTestAzAccount
-            $existingInitiatives = @(Get-AzPolicySetDefinition | Where-Object { ($_.Properties.PSObject.Properties.Name -contains 'displayName') -and ($_.Properties.displayName.Contains($newGCPolicyParameters.DisplayName) ) } )
-            $null -ne $existingInitiatives | Should -BeTrue
-            $existingInitiatives.Count | Should -Be 1
+            $existingPolicies.Count | Should -Be 1
         }
     }  
     AfterAll {
         if ($ReleaseBuild) {
             Login-ToTestAzAccount
             # Cleanup
-            $existingInitiatives = @(Get-AzPolicySetDefinition | Where-Object { ($_.Properties.PSObject.Properties.Name -contains 'displayName') -and ($_.Properties.displayName.Contains($newGCPolicyParameters.DisplayName) ) } )
-
-            foreach ($existingInitiative in $existingInitiatives) {
-                $null = Remove-AzPolicySetDefinition -Name $existingInitiative.Name -Force
+            
+            $existingPolicy = @(Get-AzPolicyDefinition | Where-Object { ($_.Properties.PSObject.Properties.Name -contains 'displayName') -and ($_.Properties.displayName.Contains($newGCPolicyParameters.DisplayName) ) } )
+            if ($null -ne $existingPolicy) {
+                $null = Remove-AzPolicyDefinition -Name $existingPolicy.Name -Force
             }
 
-            foreach ($existingPolicy in $existingPolicies) {
-                $null = Remove-AzPolicyDefinition -Name $existingPolicy.Name -Force
+            $RG = Get-AzResourceGroup "GC_Module_$DateStamp" -ErrorAction SilentlyContinue
+            if ($null -ne $RG) {
+                $null = Remove-AzResourceGroup "GC_Module_$DateStamp" -Force
             }
         }
     }