+++ helpers/GuestConfigPath.psm1
@@ -72,7 +72,7 @@ function Get-DscLibPath
 {
     $gcBinPath = Get-GuestConfigBinaryPath
     if($(Get-OSPlatform) -eq 'Windows') {
-        return Join-Path $gcBinPath 'dsclib.dll'
+        return Join-Path $gcBinPath 'gclib.dll'
     }
 
     return Join-Path $gcBinPath 'libdsclib.so'